[#139] Added first thread barriers around Nodes
