+++ clients/java-client-lib/client-core/src/main/java/net/zscript/javaclient/nodes/Connection.java
@@ -4,6 +4,7 @@ import java.util.function.Consumer;
 
 import net.zscript.javaclient.addressing.AddressedCommand;
 import net.zscript.javaclient.addressing.AddressedResponse;
+import net.zscript.javaclient.threading.ZscriptWorkerThread;
 
 public interface Connection {
     void send(AddressedCommand cmd);
@@ -11,4 +12,6 @@ public interface Connection {
     void onReceive(Consumer<AddressedResponse> resp);
 
     void responseReceived(AddressedCommand found);
+
+    ZscriptWorkerThread getAssociatedThread();
 }
